+++ src/payloads/features/idpay/get-iban-list.ts
@@ -1,0 +1,6 @@
+import { IbanListDTO } from "../../../../generated/definitions/idpay/IbanListDTO";
+import { ibanList } from "../../../persistence/idpay";
+
+export const getIbanListResponse = (): IbanListDTO => ({
+  ibanList
+});
