+++ src/payloads/features/idpay/get-instrument-list.ts
@@ -1,14 +1,14 @@
 import * as O from "fp-ts/lib/Option";
 import { pipe } from "fp-ts/lib/function";
-import { InstrumentListDTO } from "../../../../../generated/definitions/idpay/InstrumentListDTO";
-import { getInitiativeInstruments, getInstruments } from "./data";
-import { InstrumentDTO } from "../../../../../generated/definitions/idpay/InstrumentDTO";
+import { InstrumentListDTO } from "../../../../generated/definitions/idpay/InstrumentListDTO";
+import { InstrumentDTO } from "../../../../generated/definitions/idpay/InstrumentDTO";
+import { instruments } from "../../../persistence/idpay";
 
 export const getInstrumentListResponse = (
-  id: string
+  initiativeId: string
 ): O.Option<InstrumentListDTO> =>
   pipe(
-    getInitiativeInstruments(id),
+    instruments[initiativeId],
     O.fromNullable,
     O.alt(() => O.some([] as ReadonlyArray<InstrumentDTO>)),
     O.map(instrumentList => ({ instrumentList }))
