+++ src/payloads/features/idpay/get-timeline-detail.ts
@@ -1,14 +1,14 @@
 import * as O from "fp-ts/lib/Option";
 import { pipe } from "fp-ts/lib/function";
-import { OperationDTO } from "../../../../../generated/definitions/idpay/OperationDTO";
-import { getInitiativeTimelineDetails } from "./data";
+import { OperationDTO } from "../../../../generated/definitions/idpay/OperationDTO";
+import { initiativeTimelineDetails } from "../../../persistence/idpay";
 
 export const getTimelineDetailResponse = (
   initiativeId: string,
   operationId: string
 ): O.Option<OperationDTO> =>
   pipe(
-    getInitiativeTimelineDetails(initiativeId),
+    initiativeTimelineDetails[initiativeId],
     O.fromNullable,
     O.chain(details =>
       pipe(
