+++ src/payloads/features/idpay/get-timeline.ts
@@ -1,9 +1,9 @@
 import { faker } from "@faker-js/faker/locale/it";
 import * as O from "fp-ts/lib/Option";
-import _ from "lodash";
-import { TimelineDTO } from "../../../../../generated/definitions/idpay/TimelineDTO";
-import { getInitiativeTimeline } from "./data";
 import { pipe } from "fp-ts/lib/function";
+import _ from "lodash";
+import { TimelineDTO } from "../../../../generated/definitions/idpay/TimelineDTO";
+import { initiativeTimeline } from "../../../persistence/idpay";
 
 export const getTimelineResponse = (
   initiativeId: string,
@@ -11,7 +11,7 @@ export const getTimelineResponse = (
   pageSize: number = 3
 ): O.Option<TimelineDTO> =>
   pipe(
-    getInitiativeTimeline(initiativeId),
+    initiativeTimeline[initiativeId],
     O.fromNullable,
     O.map(timeline => {
       const totalElements = timeline.length;
