+++ src/payloads/features/idpay/get-wallet-detail.ts
@@ -1,0 +1,7 @@
+import * as O from "fp-ts/lib/Option";
+import { InitiativeDTO } from "../../../../generated/definitions/idpay/InitiativeDTO";
+import { initiatives } from "../../../persistence/idpay";
+
+export const getWalletDetailResponse = (
+  initiativeId: string
+): O.Option<InitiativeDTO> => O.fromNullable(initiatives[initiativeId]);
