+++ src/payloads/features/idpay/get-wallet-status.ts
@@ -1,13 +1,13 @@
 import * as O from "fp-ts/lib/Option";
 import { pipe } from "fp-ts/lib/function";
-import { WalletStatusDTO } from "../../../../../generated/definitions/idpay/WalletStatusDTO";
-import { getInitiative } from "./data";
+import { WalletStatusDTO } from "../../../../generated/definitions/idpay/WalletStatusDTO";
+import { initiatives } from "../../../persistence/idpay";
 
 export const getWalletStatusResponse = (
-  id: string
+  initiativeId: string
 ): O.Option<WalletStatusDTO> =>
   pipe(
-    getInitiative(id),
+    initiatives[initiativeId],
     O.fromNullable,
     O.map(({ status }) => ({ status }))
   );
