+++ src/payloads/features/idpay/get-wallet.ts
@@ -1,0 +1,6 @@
+import { WalletDTO } from "../../../../generated/definitions/idpay/WalletDTO";
+import { initiatives } from "../../../persistence/idpay";
+
+export const getWalletResponse = (): WalletDTO => ({
+  initiativeList: Object.values(initiatives)
+});
