--- src/payloads/features/idpay/iban/get-iban-list.ts
@@ -1,6 +1,0 @@
-import { IbanListDTO } from "../../../../../generated/definitions/idpay/IbanListDTO";
-import { getIbanList } from "./data";
-
-export const getIbanListResponse = (): IbanListDTO => ({
-  ibanList: getIbanList()
-});
