--- src/payloads/features/idpay/onboarding/get-initiative-data.ts
@@ -1,8 +1,0 @@
-import * as O from "fp-ts/lib/Option";
-import { InitiativeDataDTO } from "../../../../../generated/definitions/idpay/InitiativeDataDTO";
-import { IDPayServiceID } from "./types";
-import { initiativeData } from "./data";
-
-export const getInitiativeDataResponseByServiceId = (
-  id: IDPayServiceID
-): O.Option<InitiativeDataDTO> => O.fromNullable(initiativeData[id]);
