--- src/payloads/features/idpay/wallet/get-wallet.ts
@@ -1,6 +1,0 @@
-import { WalletDTO } from "../../../../../generated/definitions/idpay/WalletDTO";
-import { getInitiatives } from "./data";
-
-export const getWalletResponse = (): WalletDTO => ({
-  initiativeList: Object.values(getInitiatives())
-});
