+++ src/routers/features/idpay/__tests__/iban.test.ts
@@ -1,8 +1,8 @@
 import supertest from "supertest";
-import { getIbanList } from "../../../../payloads/features/idpay/iban/data";
-import { getIbanListResponse } from "../../../../payloads/features/idpay/iban/get-iban-list";
+import { getIbanListResponse } from "../../../../payloads/features/idpay/get-iban-list";
 import app from "../../../../server";
 import { addIdPayPrefix } from "../router";
+import { ibanList } from "../../../../persistence/idpay";
 
 const request = supertest(app);
 
@@ -16,7 +16,7 @@ describe("IDPay IBAN API", () => {
   });
   describe("GET getIban", () => {
     it("should return 200 with the requested IBAN if IBAN exists", async () => {
-      const iban = getIbanList()[0];
+      const iban = ibanList[0];
 
       const response = await request.get(addIdPayPrefix(`/iban/${iban?.iban}`));
       expect(response.status).toBe(200);
