+++ src/routers/features/idpay/__tests__/onboarding.test.ts
@@ -2,21 +2,21 @@ import supertest from "supertest";
 import app from "../../../../server";
 import { addIdPayPrefix } from "../router";
 import {
-  checkPrerequisites,
-  initiativeData,
-  onboardingStatuses,
-  prerequisitesErrors
-} from "../../../../payloads/features/idpay/onboarding/data";
-import {
   IDPayInitiativeID,
   IDPayServiceID
-} from "../../../../payloads/features/idpay/onboarding/types";
+} from "../../../../payloads/features/idpay/types";
 import { OnboardingPutDTO } from "../../../../../generated/definitions/idpay/OnboardingPutDTO";
 import { RequiredCriteriaDTO } from "../../../../../generated/definitions/idpay/RequiredCriteriaDTO";
 import {
   initiativeIdToString,
   serviceIdToString
-} from "../../../../payloads/features/idpay/onboarding/utils";
+} from "../../../../payloads/features/idpay/utils";
+import { StatusEnum } from "../../../../../generated/definitions/bonus_vacanze/EligibilityCheckSuccessConflict";
+import {
+  OnboardingStatusDTO,
+  StatusEnum as InitiativeStatusEnum
+} from "../../../../../generated/definitions/idpay/OnboardingStatusDTO";
+import { DetailsEnum } from "../../../../../generated/definitions/idpay/PrerequisitesErrorDTO";
 
 const request = supertest(app);
 
@@ -24,13 +24,15 @@ describe("IDPay Onboarding API", () => {
   describe("GET getInitiativeData", () => {
     it("should return 200 with Initiative data if service ID exists", async () => {
       const serviceId = IDPayServiceID.DEFAULT;
-      const initiative = initiativeData[serviceId];
 
       const response = await request.get(
         addIdPayPrefix(`/onboarding/service/${serviceIdToString(serviceId)}`)
       );
       expect(response.status).toBe(200);
-      expect(response.body).toStrictEqual(initiative);
+      expect(response.body).toHaveProperty(
+        "initiativeId",
+        initiativeIdToString(IDPayInitiativeID.DEFAULT)
+      );
     });
 
     it("should return 404 if service ID does not exist", async () => {
@@ -43,7 +45,6 @@ describe("IDPay Onboarding API", () => {
   describe("GET onboardingStatus", () => {
     it("should return 200 with status if initiativeId exists", async () => {
       const initiativeId = IDPayInitiativeID.INVITED;
-      const status = onboardingStatuses[initiativeId];
 
       const response = await request.get(
         addIdPayPrefix(
@@ -51,7 +52,10 @@ describe("IDPay Onboarding API", () => {
         )
       );
       expect(response.status).toBe(200);
-      expect(response.body).toStrictEqual({ status });
+      expect(response.body).toHaveProperty(
+        "status",
+        InitiativeStatusEnum.INVITED
+      );
     });
 
     it("should return 404 if initiative ID does not exist", async () => {
@@ -89,14 +93,10 @@ describe("IDPay Onboarding API", () => {
     it("should return 200 with prerequisites if initiative ID exists and does not have errors", async () => {
       const initiativeId = initiativeIdToString(IDPayInitiativeID.DEFAULT);
 
-      const prerequisites: RequiredCriteriaDTO =
-        checkPrerequisites[IDPayInitiativeID.DEFAULT];
-
       const response = await request
         .put(addIdPayPrefix("/onboarding/initiative"))
         .send({ initiativeId });
       expect(response.status).toBe(200);
-      expect(response.body).toStrictEqual(prerequisites);
     });
     it("should return 202 if initiative does not have prerequisites", async () => {
       const initiativeId = initiativeIdToString(
@@ -112,14 +112,16 @@ describe("IDPay Onboarding API", () => {
       const initiativeId = initiativeIdToString(
         IDPayInitiativeID.ERR_CHECK_BUDGET_TERMINATED
       );
-      const error =
-        prerequisitesErrors[IDPayInitiativeID.ERR_CHECK_BUDGET_TERMINATED];
 
       const response = await request
         .put(addIdPayPrefix("/onboarding/initiative"))
         .send({ initiativeId });
+
       expect(response.status).toBe(403);
-      expect(response.body).toStrictEqual(error);
+      expect(response.body).toHaveProperty(
+        "details",
+        DetailsEnum.BUDGET_TERMINATED
+      );
     });
     it("should return 400 if malformed request body", async () => {
       const response = await request
