+++ src/routers/features/idpay/onboarding.ts
@@ -5,15 +5,15 @@ import { getIdPayError } from "../../../payloads/features/idpay/error";
 import {
   getCheckPrerequisitesResponseByInitiativeId,
   getPrerequisitesErrorByInitiativeId
-} from "../../../payloads/features/idpay/onboarding/check-prerequisites";
-import { getInitiativeDataResponseByServiceId } from "../../../payloads/features/idpay/onboarding/get-initiative-data";
-import { getOnboardingStatusResponseByInitiativeId } from "../../../payloads/features/idpay/onboarding/onboarding-status";
+} from "../../../payloads/features/idpay/check-prerequisites";
+import { getInitiativeDataResponseByServiceId } from "../../../payloads/features/idpay/get-initiative-data";
+import { getOnboardingStatusResponseByInitiativeId } from "../../../payloads/features/idpay/onboarding-status";
 
 import { addIdPayHandler } from "./router";
 import {
   initiativeIdFromString,
   serviceIdFromString
-} from "../../../payloads/features/idpay/onboarding/utils";
+} from "../../../payloads/features/idpay/utils";
 
 /**
  * Retrieves the initiative ID starting from the corresponding service ID
